pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bluish"
version = "1.0.0"
authors = [
  { name="Luis Medel", email="luis@luismedel.com" },
]
description = "A CI/CD tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv==1.0.1",
  "pyyaml==6.0.2",
  "click==8.1.7",
]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Typing :: Typed",
]

[project.urls]
"Homepage" = "https://github.com/luismedel/bluish"

[tool.setuptools.package-data]
tsidpy = ["py.typed"]

[project.scripts]
bluish = "bluish.app:main"
